requirements.txt
Django~=4.2.1
djangorestframework~=3.14.0
djangorestframework-jwt~=1.11.0